Move the classes representing network PDUs into a separate package
