+++ client/src/main/java/replicant/Connector.java
@@ -17,6 +17,11 @@ import java.util.stream.Stream;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import org.realityforge.anodoc.TestOnly;
+import replicant.messages.ChangeSet;
+import replicant.messages.ChannelChange;
+import replicant.messages.EntityChange;
+import replicant.messages.EntityChangeData;
+import replicant.messages.EntityChannel;
 import replicant.spy.ConnectFailureEvent;
 import replicant.spy.ConnectedEvent;
 import replicant.spy.DataLoadStatus;
