+++ client/src/main/java/replicant/MessageResponse.java
@@ -5,6 +5,8 @@ import java.util.Objects;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import org.realityforge.anodoc.TestOnly;
+import replicant.messages.ChangeSet;
+import replicant.messages.EntityChange;
 import replicant.spy.DataLoadStatus;
 import static org.realityforge.braincheck.Guards.*;
