+++ client/src/main/java/replicant/messages/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Classes representing change messages passed across a Transport.
+ */
+package replicant.messages;
