+++ client/src/test/java/replicant/ChannelChangeTest.java
@@ -2,6 +2,7 @@ package replicant;
 
 import org.realityforge.guiceyloops.shared.ValueUtil;
 import org.testng.annotations.Test;
+import replicant.messages.ChannelChange;
 import static org.testng.Assert.*;
 
 public class ChannelChangeTest
