+++ client/src/test/java/replicant/ConnectorTest.java
@@ -12,6 +12,11 @@ import javax.annotation.Nullable;
 import org.intellij.lang.annotations.Language;
 import org.realityforge.guiceyloops.shared.ValueUtil;
 import org.testng.annotations.Test;
+import replicant.messages.ChangeSet;
+import replicant.messages.ChannelChange;
+import replicant.messages.EntityChange;
+import replicant.messages.EntityChangeData;
+import replicant.messages.EntityChannel;
 import replicant.spy.AreaOfInterestStatusUpdatedEvent;
 import replicant.spy.ConnectFailureEvent;
 import replicant.spy.ConnectedEvent;
